requirements.txt
# python
celery 
redis 
Pillow 
transliterate 
pandas 
bs4 
selenium 
xlsxwriter
google-api-python-client
python-magic




# django 
django

#file 
django-filer 
# django-filebrowser
django-filebrowser-no-grappelli
django-filemanager
django-bfm 







# admin-styling 
django-grappelli 
django-admin-tools 
django-jet 
# django-suit 
# https://github.com/darklow/django-suit/tarball/develop
# помилки лізуть з всіх дир
# django-material
#  не зовсім для адмінки 
# django-bootstrapped
# ....







# live settings 
django-des 
django-constance
django-solo




# other 

django-modeltranslation
django-rosetta

django-taggit
django-select2
django-allauth
django-ckeditor
django-filter
django-tinymce4-lite 
django-mptt
django-debug-toolbar
django-extensions
django-rest-framework
django-crispy-forms
django-import-export
django-decouple 
django-autocompolete-light 

django-admin-autocomplete-filter


django-celery-beat 
django-redis
channels
channels_redis 



